refactor: extract question form into standalone QuestionFormComponent
Creates app-question-form with its own template, styles, and state:
- Owns the question signal internally
- Emits submitted(question: string) and loadClicked()
- Encapsulates header, textarea, form actions, and privacy note

CouncilComponent now imports QuestionFormComponent instead of FormsModule
and delegates idle-state rendering entirely to the new component.
Removes question signal, updateQuestion(), and onSubmit() from council.
