+++ frontend/src/app/components/council/council.component.html
@@ -1,44 +1,9 @@
 <!-- Idle State -->
 @if (state().phase === 'idle') {
-  <div class="council-idle">
-    <div class="header">
-      <h1>The Sovereign Council</h1>
-      <p>A privacy-first council of local LLMs ready to deliberate</p>
-    </div>
-
-    <form class="question-form" (ngSubmit)="onSubmit()">
-      <textarea
-        class="input"
-        [ngModel]="question()"
-        (ngModelChange)="updateQuestion($event)"
-        name="question"
-        placeholder="What question would you like the council to deliberate?"
-        rows="4"
-      ></textarea>
-
-      <div class="form-actions">
-        <button
-          type="submit"
-          class="btn btn-primary flex-1"
-          [disabled]="!question().trim()"
-        >
-          Convene the Council
-        </button>
-        <button
-          type="button"
-          class="btn btn-secondary"
-          (click)="onLoadClick()"
-        >
-          Load Saved
-        </button>
-      </div>
-    </form>
-
-    <div class="privacy-note">
-      <p>All processing happens locally on your machine.</p>
-      <p>Your questions never leave your device.</p>
-    </div>
-  </div>
+  <app-question-form
+    (submitted)="onQuestionSubmitted($event)"
+    (loadClicked)="onLoadClick()"
+  />
 }
 
 <!-- Loading State -->
