+++ frontend/src/app/components/council/council.component.ts
@@ -1,17 +1,17 @@
-import { Component, signal, computed, inject } from '@angular/core';
-import { FormsModule } from '@angular/forms';
+import { Component, computed, inject, signal } from '@angular/core';
 import { toSignal } from '@angular/core/rxjs-interop';
 import { DeliberationService } from '../../services/deliberation.service';
 import { DeliberationViewComponent } from '../deliberation-view/deliberation-view.component';
 import { ErrorStateComponent } from '../error-state/error-state.component';
 import { LoadingStateComponent } from '../loading-state/loading-state.component';
 import { SaveLoadDialogComponent } from '../save-load-dialog/save-load-dialog.component';
+import { QuestionFormComponent } from '../question-form/question-form.component';
 
 @Component({
   selector: 'app-council',
   standalone: true,
   imports: [
-    FormsModule,
+    QuestionFormComponent,
     DeliberationViewComponent,
     ErrorStateComponent,
     LoadingStateComponent,
@@ -23,31 +23,19 @@ import { SaveLoadDialogComponent } from '../save-load-dialog/save-load-dialog.co
 export class CouncilComponent {
   private readonly deliberationService = inject(DeliberationService);
 
-  // Local state signals
-  readonly question = signal('');
   readonly dialogMode = signal<'save' | 'load' | null>(null);
 
-  // Convert service observable to signal
   readonly state = toSignal(this.deliberationService.state$, {
     initialValue: this.deliberationService.state,
   });
 
-  // Computed signals
   readonly isLoading = computed(() => {
     const phase = this.state().phase;
     return ['gathering', 'reviewing', 'synthesizing', 'analyzing'].includes(phase);
   });
 
-  onSubmit(): void {
-    const q = this.question().trim();
-    if (q && !this.isLoading()) {
-      this.deliberationService.ask(q);
-      this.question.set('');
-    }
-  }
-
-  updateQuestion(value: string): void {
-    this.question.set(value);
+  onQuestionSubmitted(question: string): void {
+    this.deliberationService.ask(question);
   }
 
   onSave(): void {
@@ -65,7 +53,6 @@ export class CouncilComponent {
   onCancel(): void {
     if (confirm('Are you sure you want to cancel this deliberation?')) {
       this.deliberationService.cancel();
-      this.question.set(''); // Clear question input
     }
   }
