+++ frontend/src/app/components/question-form/question-form.component.scss
@@ -1,0 +1,55 @@
+.question-form-root {
+  max-width: 32rem;
+  margin: 0 auto;
+}
+
+.header {
+  text-align: center;
+  margin-bottom: 2rem;
+
+  h1 {
+    font-size: var(--font-size-h2);
+    font-weight: var(--font-weight-bold);
+    color: var(--color-text-default);
+    margin: 0 0 0.5rem;
+  }
+
+  p {
+    color: var(--color-text-subtle-on-surface);
+    margin: 0;
+  }
+}
+
+.question-form {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+
+  textarea {
+    min-height: 6rem;
+  }
+}
+
+.form-actions {
+  display: flex;
+  gap: 0.75rem;
+
+  .flex-1 {
+    flex: 1;
+  }
+}
+
+.privacy-note {
+  margin-top: 2rem;
+  text-align: center;
+
+  p {
+    font-size: var(--font-size-sm);
+    color: var(--color-text-subtle);
+    margin: 0;
+
+    & + p {
+      margin-top: 0.25rem;
+    }
+  }
+}
